requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
social-auth-app-django
python-decouple
pyotp
django-celery